refactor Matcher and FlavorSpec
Scoring logic should live in the FlavorSpec and provide an easy way to
obtain score for a given Machine.

This reduces Matcher functionality to simply interacting with a
collection of flavors and matching them against Machine.
